refactor(client): extract TempoClient to src/client/ for multi-interface reuse
Move TempoClient interface, types, and factory from src/tui/client.ts to
src/client/ so non-TUI consumers (CLI, tests, integrations) can depend on
it without pulling in Ink/React. The old tui/client.ts becomes a thin
re-export for backward compatibility.

Also replaces dead ClaudeTempoIsConductor search attribute reads in the
fallback paths with workflow ID convention detection.

Closes #68
